fix bug with Task 0
Previously, Task 0 was not being handled correctly. The worker that was
assigned Task 0 would not receive a task with 0 index in the task
assignment reply. Instead, the task index would be the same as whatever
previous task it had been working on. This was because we were using a
single, common TaskInfo struct for each RPC. Because 0 is the zero value
for ints, I presume gob simply ignored that particular index when
decoding, thus keeping the previous index.