Factor out MR execution logic into mapreduce.go
Previously, we started the Coordinator after starting all the Workers in
goroutines. This opened ourselves up to a race condition where it was
possible (in theory) that the Workers would begin sending RPC requests
to Coordinator that had not yet started and crash. To fix this, we have
to start the Coordinator before the Workers, but this requires some
reorganization.

We move the code that actually executes the Coordinator and Workers out
from the various places it was spread out before (run-mr.go,
coordinator.go and worker.go) into a dedicated RunMapReduce function in
mapreduce.go. This consolidates the execution logic in one place. It
also gives us more control over the other in which actions happen.