Partition seq output to make testing easier
Partition sequential output like we do the parallel output so that
testing equivalence between them is easier. Previously, the way that
Go sorted strings with Unicode characters (in sequential output) and the
way that the sort utility sorted strings with Unicode characters (we
used sort to combine parallel outputs) would differ.